Split services/hasheous into a client and transport module
The deadline-aware HTTPS stack is a self-contained concern -- its own
socket classes, its own connection and handler, its own failure modes --
and it had grown to roughly a third of the file. It moves to
services/hasheous/transport.py, leaving the client (toggle, cooldown,
_fetch_json, lookup, health, normalization) in __init__.py.

No behaviour change: the package re-exports the transport names, so
`from services import hasheous` still reaches the whole surface and
_fetch_json stays the single patch seam. That matters more than it
looks -- a second module object for this client is exactly how a test
suite ends up silently hitting the real network, and a guard test
already exists for that.
